Move out swagger configuration
